backend/recognize/requirements.txt
functions-framework==3.8.2
google-cloud-vision==3.13.0
requests==2.32.5
